Tweak to Smooth Movement
+ gets the Turn Time and uses that as Move Time.
++ Needs to be in Start90 because Start() is executed after
GameController.Awake() (which initializes the GameController.instance)

PS: The door disappearing is not a bug. There is a lever at one of the
tiles, which is linked to that door. It is exactly what should happen BUT
the lever is TINY as f*ck and you probably did not see it.