Major refactoring; Spawns; Tags; LevelManager
+ The LevelManager script has been tidied!
++ There are now labeled sections for the script's public variables.
++ The prefab objects (Monkey, Ghost) are assigned to the script itself
(before even being added as a component on a game object)

+ A new Tile called "Spawn" has been created
++ Spawn tiles denote the spawn points on the level. This replaces the old
list on the LevelManager script component. It is better because it even
gives visual feedback in the scene view.
++ A corresponding 'Spawn' script was also created in Scripts/Items, which
allows the Spawn Points to be ordered however the level designer chooses.
This is achieved by specifying a 'spawn index' for each spawn.

+ The Tags have been refactored.
++ Unused tags such as 'Banana' and 'Bomb' were removed
++ New tags were introduced for 'Spawn' and 'Exit'
++ The LevelManager now uses these new tags to find the Tiles it needs.
This removes the need for the public list variable on the script component

+ The Scene has been saved to reflect the usage of the new tags