Fixed a Banana bug; Changed Collider check method
+ The Banana Script now does not check distance to the other collider to
decide whether the inner collider was activated or the outer one. Instead,
the inner collider is referenced now, and a proper check is done if IT was
the one that triggered the event.
++ Finding out which collider was the trigger is important to distinguish
between the two effects: Pick Up or Lure Monkey
+++ If there was no distinction, either the player would be able to pick
the Banana from units away, OR the monkey would have to go on top of the
banana to be lured to it. Neither of those cases are acceptable